Gate user calls with call-target evidence
