+++ crates/nose-normalize/src/value_graph.rs
@@ -35,7 +35,12 @@
 //! proof-obligation: normalize.value_graph.functor
 //! proof-obligation: normalize.value_graph.min_max
 
+mod context;
+mod ops;
 mod rules;
+mod stdlib;
+
+pub use context::ValueFingerprintContext;
 
 use crate::combine;
 use crate::module_facts::{
@@ -85,6 +90,7 @@ use nose_semantics::{
     StaticIndexMembershipKind, ValueDomain, ValueLaw, SEQ_VALUE_COLLECTION, SEQ_VALUE_MAP,
     SEQ_VALUE_OWN_PROPERTY_GUARD, SEQ_VALUE_PAIR, SEQ_VALUE_RECORD_GUARD, SEQ_VALUE_UNTAGGED,
 };
+use ops::*;
 use rustc_hash::{FxHashMap, FxHashSet};
 use std::borrow::Cow;
 use std::sync::OnceLock;
@@ -168,146 +174,6 @@ pub fn value_fingerprint_lits_anchors_with_context(
     let (v, l, r) = b.fingerprint_lits();
     let a = b.anchors(ANCHOR_MIN_WEIGHT);
     (v, l, r, a)
-}
-
-/// File-level facts that are independent of the unit currently being fingerprinted.
-///
-/// `units::extract` may fingerprint hundreds of block units from the same large file.
-/// Function binding proofs require scanning every function and building a
-/// literal-sensitive subtree hash for the whole IL, and opaque raw/lambda values need a
-/// structural subtree hash for the same file. Doing either once per unit turns a
-/// file-level proof into the dominant cost. This context keeps the reusable proof result
-/// and lazily shares structural subtree hashes. Each per-unit builder still interns
-/// the corresponding lambda values into its own value arena, so value ids never cross
-/// builder boundaries.
-pub struct ValueFingerprintContext {
-    module: ModuleSeedContext,
-    function_bindings: Vec<(Symbol, u64)>,
-    subtree_hashes: OnceLock<Vec<u64>>,
-}
-
-impl ValueFingerprintContext {
-    pub fn new(il: &Il, interner: &Interner) -> Self {
-        let module = ModuleSeedContext::new(il, interner);
-        let subtree_hashes = OnceLock::new();
-        let function_bindings = {
-            let mut b = Builder::new(il, interner)
-                .with_shared_subtree_hashes(&subtree_hashes)
-                .with_local_scope_nodes(&module.local_scope);
-            b.seed_module_value_bindings_from_context(&module, None);
-            b.collect_function_binding_hashes()
-        };
-        Self {
-            module,
-            function_bindings,
-            subtree_hashes,
-        }
-    }
-}
-
-struct ModuleSeedContext {
-    local_scope: Vec<bool>,
-    top_level: Vec<NodeId>,
-    assignment_counts: FxHashMap<Symbol, usize>,
-    assignment_deps: FxHashMap<Symbol, FxHashSet<Symbol>>,
-    mutated_bindings: FxHashSet<Symbol>,
-    unit_symbols: FxHashSet<Symbol>,
-}
-
-impl ModuleSeedContext {
-    fn new(il: &Il, interner: &Interner) -> Self {
-        let local_scope = local_scope_nodes(il);
-        let top_level = top_level_statements_for(il);
-        let mut is_top_level = vec![false; il.nodes.len()];
-        for &stmt in &top_level {
-            if let Some(slot) = is_top_level.get_mut(stmt.0 as usize) {
-                *slot = true;
-            }
-        }
-
-        let mut assignment_counts: FxHashMap<Symbol, usize> = FxHashMap::default();
-        for &stmt in &top_level {
-            if let Some(name) = module_seed_assignment_name(il, stmt, &local_scope) {
-                *assignment_counts.entry(name).or_insert(0) += 1;
-            }
-        }
-        let mut assignment_deps: FxHashMap<Symbol, FxHashSet<Symbol>> = FxHashMap::default();
-        for &stmt in &top_level {
-            let Some(name) = module_seed_assignment_name(il, stmt, &local_scope) else {
-                continue;
-            };
-            if let Some(&rhs) = il.children(stmt).get(1) {
-                let mut deps = FxHashSet::default();
-                collect_all_node_symbols_in_scope(il, rhs, &local_scope, &mut deps);
-                assignment_deps.insert(name, deps);
-            }
-        }
-
-        let unit_symbols: FxHashSet<Symbol> =
-            il.units.iter().filter_map(|unit| unit.name).collect();
-        let candidate_names: FxHashSet<Symbol> = assignment_counts
-            .iter()
-            .filter_map(|(&name, &count)| {
-                (count == 1 && !unit_symbols.contains(&name)).then_some(name)
-            })
-            .collect();
-        let direct_definitions: FxHashSet<NodeId> = top_level
-            .iter()
-            .copied()
-            .filter(|&stmt| module_seed_assignment_name(il, stmt, &local_scope).is_some())
-            .collect();
-        let mutated_bindings = collect_module_mutations_in_scope_with_direct_definitions(
-            il,
-            interner,
-            &candidate_names,
-            &is_top_level,
-            &local_scope,
-            &direct_definitions,
-        );
-
-        Self {
-            local_scope,
-            top_level,
-            assignment_counts,
-            assignment_deps,
-            mutated_bindings,
-            unit_symbols,
-        }
-    }
-
-    fn required_bindings_for(&self, il: &Il, root: NodeId) -> FxHashSet<Symbol> {
-        let mut required = FxHashSet::default();
-        collect_all_node_symbols_in_scope(il, root, &self.local_scope, &mut required);
-        let mut stack: Vec<Symbol> = required.iter().copied().collect();
-        while let Some(name) = stack.pop() {
-            let Some(deps) = self.assignment_deps.get(&name) else {
-                continue;
-            };
-            for &dep in deps {
-                if self.assignment_counts.contains_key(&dep) && required.insert(dep) {
-                    stack.push(dep);
-                }
-            }
-        }
-        required
-    }
-}
-
-fn module_seed_assignment_name(il: &Il, stmt: NodeId, local_scope: &[bool]) -> Option<Symbol> {
-    assignment_name_in_scope(il, stmt, local_scope)
-        .or_else(|| evidence_backed_raw_assignment_name(il, stmt))
-}
-
-fn evidence_backed_raw_assignment_name(il: &Il, stmt: NodeId) -> Option<Symbol> {
-    let (lhs, rhs) = il.assignment_var_parts(stmt)?;
-    let symbol = il.var_name(lhs)?;
-    if import_fact_evidence_rhs(il, rhs).is_some()
-        || imported_literal_producer_evidence_for_node(il, rhs)
-    {
-        Some(symbol)
-    } else {
-        None
-    }
 }
 
 pub fn value_fingerprint_lits_with_context(
@@ -608,21 +474,6 @@ impl<'a> Builder<'a> {
             clamp_candidate_count: 0,
             clamp_proof_backed_candidate_count: 0,
         }
-    }
-
-    fn with_shared_subtree_hashes(mut self, hashes: &'a OnceLock<Vec<u64>>) -> Self {
-        self.shared_subtree_hashes = Some(hashes);
-        self
-    }
-
-    fn with_local_scope_nodes(mut self, local_scope_nodes: &'a [bool]) -> Self {
-        self.local_scope_nodes = Cow::Borrowed(local_scope_nodes);
-        self
-    }
-
-    fn with_context(self, context: &'a ValueFingerprintContext) -> Self {
-        self.with_shared_subtree_hashes(&context.subtree_hashes)
-            .with_local_scope_nodes(&context.module.local_scope)
     }
 
     fn vty(&self, v: ValueId) -> ValueDomain {
@@ -962,1408 +813,6 @@ impl<'a> Builder<'a> {
             self.nodes[value as usize].op,
             ValOp::Input(key) if key == self.free_name_input_key(name)
         )
-    }
-
-    /// Shared skeleton of the collection-factory recognizers: a collection sequence literal
-    /// call `Call(0, [callee, Seq(collection)])`
-    /// whose `callee` passes `is_factory` wraps the sequence literal `args[1]`; return it. The
-    /// per-language recognizers differ ONLY in their callee predicate, so this collapses the
-    /// identical skeletons that nose's own duplication gate flagged across them.
-    fn collection_factory_seq(
-        &self,
-        value: ValueId,
-        is_factory: impl FnOnce(&Self, ValueId) -> bool,
-    ) -> Option<ValueId> {
-        let node = &self.nodes[value as usize];
-        if !matches!(node.op, ValOp::Call(0)) || node.args.len() != 2 {
-            return None;
-        }
-        let (callee, seq) = (node.args[0], node.args[1]);
-        if !matches!(
-            self.nodes[seq as usize].op,
-            ValOp::Seq(SEQ_VALUE_COLLECTION)
-        ) {
-            return None;
-        }
-        is_factory(self, callee).then_some(seq)
-    }
-
-    /// `factory(<seq>)` where `factory` is a free function/path name that constructs a collection
-    /// from a single sequence literal. Data-driven by the first-party collection contracts in
-    /// `nose-semantics`; each row carries the names and whether a same-named local definition
-    /// shadows the builtin.
-    fn proven_free_name_collection_factory(&self, value: ValueId) -> Option<ValueId> {
-        self.collection_factory_seq(value, |s, callee| {
-            let node = &s.nodes[callee as usize];
-            let ValOp::Input(key) = node.op else {
-                return false;
-            };
-            let Some(contract) = library_free_name_collection_factory_contracts(s.il.meta.lang)
-                .find(|contract| {
-                    let LibraryApiCalleeContract::FreeName { name, .. } = contract.callee else {
-                        return false;
-                    };
-                    key == s.free_name_input_key(name)
-                })
-            else {
-                return false;
-            };
-            let LibraryApiCalleeContract::FreeName { name, .. } = contract.callee else {
-                return false;
-            };
-            s.is_free_name_value(callee, name)
-                && matches!(
-                    s.library_api_evidence_for_value_call(
-                        value,
-                        callee,
-                        None,
-                        contract.id,
-                        contract.callee,
-                        1,
-                    ),
-                    LibraryApiEvidenceStatus::Admitted
-                )
-        })
-    }
-
-    /// Python `from collections import deque; deque(<seq>)` — the imported-stdlib collection
-    /// factory (the non-free-name part of the former python recognizer).
-    fn proven_python_deque_collection_value(&self, value: ValueId) -> Option<ValueId> {
-        self.collection_factory_seq(value, |s, callee| {
-            library_imported_collection_factory_contracts(s.il.meta.lang).any(|contract| {
-                let LibraryApiCalleeContract::ImportedBinding { .. } = contract.callee else {
-                    return false;
-                };
-                let receiver = match s.nodes[callee as usize].op {
-                    ValOp::Field(_) => s.nodes[callee as usize].args.first().copied(),
-                    _ => None,
-                };
-                match s.library_api_evidence_for_value_call(
-                    value,
-                    callee,
-                    receiver,
-                    contract.id,
-                    contract.callee,
-                    1,
-                ) {
-                    LibraryApiEvidenceStatus::Admitted => true,
-                    LibraryApiEvidenceStatus::Rejected => false,
-                    LibraryApiEvidenceStatus::Missing => false,
-                }
-            })
-        })
-    }
-
-    fn proven_java_collection_factory_value(&mut self, value: ValueId) -> Option<ValueId> {
-        if !semantics(self.il.meta.lang)
-            .stdlib()
-            .java_collection_factories()
-        {
-            return None;
-        }
-        let node = &self.nodes[value as usize];
-        if !matches!(node.op, ValOp::Call(0)) || node.args.len() < 2 {
-            return None;
-        }
-        let args = node.args.clone();
-        let callee = &self.nodes[args[0] as usize];
-        let ValOp::Field(method) = callee.op else {
-            return None;
-        };
-        if callee.args.len() != 1 {
-            return None;
-        }
-        let receiver = callee.args[0];
-        let contract = ["List", "Set", "Arrays"]
-            .into_iter()
-            .find_map(|receiver_name| {
-                let contract = library_java_collection_factory_contract_by_hash(
-                    self.il.meta.lang,
-                    receiver_name,
-                    method,
-                )?;
-                let LibraryApiCalleeContract::JavaUtilStaticMember { .. } = contract.callee else {
-                    return None;
-                };
-                match self.library_api_evidence_for_value_call(
-                    value,
-                    args[0],
-                    Some(receiver),
-                    contract.id,
-                    contract.callee,
-                    args.len().saturating_sub(1),
-                ) {
-                    LibraryApiEvidenceStatus::Admitted => Some(contract),
-                    LibraryApiEvidenceStatus::Rejected => None,
-                    LibraryApiEvidenceStatus::Missing => None,
-                }
-            })?;
-        // A single argument to a varargs collection factory (`Arrays.asList(x)`,
-        // `List.of(x)`, `Set.of(x)`) is ambiguous: when `x` is an array it is spread
-        // into the element list, but when `x` is a single object it is the sole
-        // element. The two readings have different membership semantics
-        // (`value` in the array elements vs `value.equals(x)`), so a single argument
-        // can only be canonicalized when the receiver is a proven array. Otherwise we
-        // must refuse, or an array-typed field and a list-typed field of the same name
-        // would false-merge. Multi-argument factories are always a literal element list.
-        if args.len() == 2 {
-            let single_arg_spreads_array = match contract.result {
-                LibraryCollectionFactoryResult::VariadicElements {
-                    single_arg_spreads_array,
-                } => single_arg_spreads_array,
-                _ => false,
-            };
-            if single_arg_spreads_array && self.is_array_param_value(args[1]) {
-                return Some(self.mk(ValOp::ArrayParam, vec![args[1]]));
-            }
-            return None;
-        }
-        Some(self.mk(ValOp::Seq(SEQ_VALUE_COLLECTION), args[1..].to_vec()))
-    }
-
-    fn eval_java_collection_constructor_expr(
-        &mut self,
-        expr: NodeId,
-        kids: &[NodeId],
-    ) -> Option<ValueId> {
-        if kids.len() != 1 || self.il.kind(kids[0]) != NodeKind::Var {
-            return None;
-        }
-        let Payload::Name(type_name) = self.il.node(kids[0]).payload else {
-            return None;
-        };
-        let contract = library_java_collection_constructor_contract(
-            self.il.meta.lang,
-            self.interner.resolve(type_name),
-            kids.len().saturating_sub(1),
-        )?;
-        let LibraryApiCalleeContract::JavaUtilConstructor { .. } = contract.callee else {
-            return None;
-        };
-        match library_api_contract_evidence_for_call(
-            self.il,
-            self.interner,
-            expr,
-            contract.id,
-            contract.callee,
-            kids.len().saturating_sub(1),
-        ) {
-            LibraryApiEvidenceStatus::Admitted => {}
-            LibraryApiEvidenceStatus::Rejected => return None,
-            LibraryApiEvidenceStatus::Missing => return None,
-        }
-        match contract.result {
-            LibraryCollectionFactoryResult::EmptySequence => {
-                Some(self.mk(ValOp::Seq(SEQ_VALUE_COLLECTION), vec![]))
-            }
-            _ => None,
-        }
-    }
-
-    fn proven_ruby_set_factory_value(&self, value: ValueId) -> Option<ValueId> {
-        self.collection_factory_seq(value, |s, callee_value| {
-            let callee = &s.nodes[callee_value as usize];
-            let ValOp::Field(method) = callee.op else {
-                return false;
-            };
-            let Some(contract) =
-                library_ruby_set_factory_contract_by_hash(s.il.meta.lang, "Set", method, 1)
-            else {
-                return false;
-            };
-            let LibraryApiCalleeContract::RubyRequireStaticMember { receiver, .. } =
-                contract.callee
-            else {
-                return false;
-            };
-            callee.args.len() == 1
-                && s.is_free_name_value(callee.args[0], receiver)
-                && matches!(
-                    s.library_api_evidence_for_value_call(
-                        value,
-                        callee_value,
-                        Some(callee.args[0]),
-                        contract.id,
-                        contract.callee,
-                        1,
-                    ),
-                    LibraryApiEvidenceStatus::Admitted
-                )
-        })
-    }
-
-    fn proven_rust_vec_macro_collection_value(&mut self, value: ValueId) -> Option<ValueId> {
-        if !semantics(self.il.meta.lang)
-            .stdlib()
-            .rust_vec_macro_factory()
-        {
-            return None;
-        }
-        let node = &self.nodes[value as usize];
-        if !matches!(node.op, ValOp::Call(0)) || node.args.is_empty() {
-            return None;
-        }
-        let args = node.args.clone();
-        let contract = library_rust_vec_macro_factory_contract(self.il.meta.lang, "vec")?;
-        let LibraryApiCalleeContract::RustMacro { name, .. } = contract.callee else {
-            return None;
-        };
-        if !self.is_free_name_value(args[0], name)
-            || !matches!(
-                self.library_api_evidence_for_value_call(
-                    value,
-                    args[0],
-                    None,
-                    contract.id,
-                    contract.callee,
-                    args.len().saturating_sub(1),
-                ),
-                LibraryApiEvidenceStatus::Admitted
-            )
-        {
-            return None;
-        }
-        Some(self.mk(ValOp::Seq(SEQ_VALUE_COLLECTION), args[1..].to_vec()))
-    }
-
-    fn is_import_namespace_expr(
-        &mut self,
-        expr: NodeId,
-        module: &str,
-        env: &FxHashMap<u32, ValueId>,
-    ) -> bool {
-        let value = self.eval(expr, env);
-        self.is_import_namespace_value(value, module)
-    }
-
-    fn is_import_namespace_value(&self, value: ValueId, module: &str) -> bool {
-        let node = &self.nodes[value as usize];
-        matches!(
-            node.op,
-            ValOp::ImportNamespace { module_hash }
-                if module_hash == stable_symbol_hash(module)
-        )
-    }
-
-    #[cfg(test)]
-    fn is_import_binding_value(&self, value: ValueId, module: &str, exported: &str) -> bool {
-        let node = &self.nodes[value as usize];
-        matches!(
-            node.op,
-            ValOp::ImportBinding {
-                module_hash,
-                exported_hash,
-            } if module_hash == stable_symbol_hash(module)
-                && exported_hash == stable_symbol_hash(exported)
-        )
-    }
-
-    fn import_fact_value(&mut self, expr: NodeId) -> Option<ValueId> {
-        let fact = import_fact_evidence_rhs(self.il, expr)?;
-        match fact.kind {
-            ImportFactKind::Namespace => Some(self.mk(
-                ValOp::ImportNamespace {
-                    module_hash: fact.module_hash,
-                },
-                vec![],
-            )),
-            ImportFactKind::Binding => Some(self.mk(
-                ValOp::ImportBinding {
-                    module_hash: fact.module_hash,
-                    exported_hash: fact.exported_hash?,
-                },
-                vec![],
-            )),
-        }
-    }
-
-    fn file_imports_namespace(&self, expr: NodeId, module: &str) -> bool {
-        semantics(self.il.meta.lang)
-            .modules()
-            .go_import_namespace_facts()
-            && imported_namespace_symbol(self.il, self.interner, expr, module)
-    }
-
-    fn proven_collection_value(&mut self, value: ValueId) -> Option<ValueId> {
-        if matches!(
-            self.nodes[value as usize].op,
-            ValOp::Seq(SEQ_VALUE_COLLECTION)
-        ) {
-            return Some(value);
-        }
-        self.proven_free_name_collection_factory(value)
-            .or_else(|| self.proven_java_collection_factory_value(value))
-            .or_else(|| self.proven_python_deque_collection_value(value))
-            .or_else(|| self.proven_ruby_set_factory_value(value))
-            .or_else(|| self.proven_rust_vec_macro_collection_value(value))
-    }
-
-    fn proven_collection_expr(
-        &mut self,
-        expr: NodeId,
-        env: &FxHashMap<u32, ValueId>,
-    ) -> Option<ValueId> {
-        let value = self.eval(expr, env);
-        self.proven_collection_value(value)
-            .or_else(|| self.proven_local_collection_binding_value(expr, env))
-    }
-
-    fn proven_local_collection_binding_value(
-        &mut self,
-        expr: NodeId,
-        env: &FxHashMap<u32, ValueId>,
-    ) -> Option<ValueId> {
-        let value = self.proven_local_binding_initializer_value(expr, env, |domain| {
-            domain.is_collection_or_set()
-        })?;
-        self.proven_collection_value(value)
-    }
-
-    fn proven_local_map_binding_value(
-        &mut self,
-        expr: NodeId,
-        env: &FxHashMap<u32, ValueId>,
-    ) -> Option<ValueId> {
-        let value =
-            self.proven_local_binding_initializer_value(expr, env, |domain| domain.is_map())?;
-        self.proven_map_value(value)
-    }
-
-    fn proven_local_binding_initializer_value(
-        &mut self,
-        expr: NodeId,
-        env: &FxHashMap<u32, ValueId>,
-        accepts_domain: impl FnOnce(DomainEvidence) -> bool,
-    ) -> Option<ValueId> {
-        if self.il.kind(expr) != NodeKind::Var {
-            return None;
-        }
-        let Payload::Cid(cid) = self.il.node(expr).payload else {
-            return None;
-        };
-        let (lhs, rhs) = self.local_binding_initializer(cid, expr)?;
-        if !nose_semantics::domain_evidence_for_binding_lhs(self.il, self.interner, lhs)
-            .is_some_and(accepts_domain)
-        {
-            return None;
-        }
-        if self.node_contains_cid(rhs, cid) {
-            return None;
-        }
-        Some(self.eval(rhs, env))
-    }
-
-    fn local_binding_initializer(&self, cid: u32, use_node: NodeId) -> Option<(NodeId, NodeId)> {
-        let mut rhs = None;
-        for (idx, node) in self.il.nodes.iter().enumerate() {
-            if node.kind != NodeKind::Assign {
-                continue;
-            }
-            let assign = NodeId(idx as u32);
-            let kids = self.il.children(assign);
-            if kids.len() != 2 {
-                continue;
-            }
-            if self.node_refers_to_cid(kids[0], cid) {
-                if node.span.end_byte > self.il.node(use_node).span.start_byte {
-                    continue;
-                }
-                if rhs.is_some() {
-                    return None;
-                }
-                rhs = Some((kids[0], kids[1]));
-            } else if self.node_contains_cid(kids[0], cid) {
-                return None;
-            }
-        }
-        rhs
-    }
-
-    /// `factory([<entry>, …])` where `factory` is a free name that builds a map from a sequence of
-    /// 2-element key/value entries. Data-driven by first-party map contracts in `nose-semantics`;
-    /// the matched row's `Seq` tag says how each entry is shaped (JS array vs Rust tuple).
-    fn proven_free_name_map_factory(&mut self, value: ValueId) -> Option<ValueId> {
-        let node = &self.nodes[value as usize];
-        if !matches!(node.op, ValOp::Call(0)) || node.args.len() != 2 {
-            return None;
-        }
-        let (callee, seq) = (node.args[0], node.args[1]);
-        if !matches!(
-            self.nodes[seq as usize].op,
-            ValOp::Seq(SEQ_VALUE_COLLECTION)
-        ) {
-            return None;
-        }
-        let entry_tag =
-            library_free_name_map_factory_contracts(self.il.meta.lang).find_map(|contract| {
-                let LibraryApiCalleeContract::FreeName { name, .. } = contract.callee else {
-                    return None;
-                };
-                if !self.is_free_name_value(callee, name)
-                    || !matches!(
-                        self.library_api_evidence_for_value_call(
-                            value,
-                            callee,
-                            None,
-                            contract.id,
-                            contract.callee,
-                            1,
-                        ),
-                        LibraryApiEvidenceStatus::Admitted
-                    )
-                {
-                    return None;
-                }
-                match contract.result {
-                    LibraryMapFactoryResult::EntrySequence { entry_seq_tag } => Some(entry_seq_tag),
-                    LibraryMapFactoryResult::JavaFactory { .. } => None,
-                }
-            })?;
-        self.map_factory_from_seq(seq, entry_tag)
-    }
-
-    /// Canonicalize a collection sequence of 2-element entries to the canonical map shape.
-    /// Shared by the free-name and (entry-wise) other map factories.
-    fn map_factory_from_seq(&mut self, seq: ValueId, entry_tag: u64) -> Option<ValueId> {
-        let entries = self.nodes[seq as usize].args.clone();
-        let mut canonical_entries = Vec::with_capacity(entries.len());
-        for entry in entries {
-            let entry_node = &self.nodes[entry as usize];
-            if !matches!(entry_node.op, ValOp::Seq(t) if t == entry_tag)
-                || entry_node.args.len() != 2
-            {
-                return None;
-            }
-            let kv = entry_node.args.clone();
-            canonical_entries.push(self.mk(ValOp::Seq(SEQ_VALUE_PAIR), kv));
-        }
-        Some(self.mk(ValOp::Seq(SEQ_VALUE_MAP), canonical_entries))
-    }
-
-    fn proven_java_map_factory_entries(&mut self, value: ValueId) -> Option<ValueId> {
-        if !semantics(self.il.meta.lang).stdlib().java_map_factories() {
-            return None;
-        }
-        let node = &self.nodes[value as usize];
-        if !matches!(node.op, ValOp::Call(0)) || node.args.is_empty() {
-            return None;
-        }
-        let args = node.args.clone();
-        let callee = &self.nodes[args[0] as usize];
-        let ValOp::Field(method) = callee.op else {
-            return None;
-        };
-        if callee.args.len() != 1 {
-            return None;
-        }
-        let contract = library_java_map_factory_contract_by_hash(self.il.meta.lang, "Map", method)?;
-        let LibraryApiCalleeContract::JavaUtilStaticMember { .. } = contract.callee else {
-            return None;
-        };
-        let api_status = self.library_api_evidence_for_value_call(
-            value,
-            args[0],
-            Some(callee.args[0]),
-            contract.id,
-            contract.callee,
-            args.len().saturating_sub(1),
-        );
-        match api_status {
-            LibraryApiEvidenceStatus::Admitted => {}
-            LibraryApiEvidenceStatus::Rejected => return None,
-            LibraryApiEvidenceStatus::Missing => return None,
-        }
-        let LibraryMapFactoryResult::JavaFactory { kind } = contract.result else {
-            return None;
-        };
-        if kind == JavaMapFactoryKind::Of {
-            let entries = &args[1..];
-            if entries.len() % 2 != 0 {
-                return None;
-            }
-            let mut canonical_entries = Vec::with_capacity(entries.len() / 2);
-            for kv in entries.chunks(2) {
-                canonical_entries.push(self.mk(ValOp::Seq(SEQ_VALUE_PAIR), kv.to_vec()));
-            }
-            return Some(self.mk(ValOp::Seq(SEQ_VALUE_MAP), canonical_entries));
-        }
-        if kind == JavaMapFactoryKind::OfEntries {
-            let mut canonical_entries = Vec::with_capacity(args.len().saturating_sub(1));
-            for entry in args.iter().skip(1).copied() {
-                let kv = self.proven_java_map_entry_pair(entry)?;
-                canonical_entries.push(self.mk(ValOp::Seq(SEQ_VALUE_PAIR), kv));
-            }
-            return Some(self.mk(ValOp::Seq(SEQ_VALUE_MAP), canonical_entries));
-        }
-        None
-    }
-
-    fn eval_java_map_factory_expr(
-        &mut self,
-        expr: NodeId,
-        kids: &[NodeId],
-        env: &FxHashMap<u32, ValueId>,
-    ) -> Option<ValueId> {
-        if !semantics(self.il.meta.lang).stdlib().java_map_factories()
-            || kids.is_empty()
-            || self.il.kind(kids[0]) != NodeKind::Field
-        {
-            return None;
-        }
-        let Payload::Name(method) = self.il.node(kids[0]).payload else {
-            return None;
-        };
-        let contract = library_java_map_factory_contract_by_hash(
-            self.il.meta.lang,
-            "Map",
-            self.interner.symbol_hash(method),
-        )?;
-        let LibraryApiCalleeContract::JavaUtilStaticMember { .. } = contract.callee else {
-            return None;
-        };
-        match library_api_contract_evidence_for_call(
-            self.il,
-            self.interner,
-            expr,
-            contract.id,
-            contract.callee,
-            kids.len().saturating_sub(1),
-        ) {
-            LibraryApiEvidenceStatus::Admitted => {}
-            LibraryApiEvidenceStatus::Rejected => return None,
-            LibraryApiEvidenceStatus::Missing => return None,
-        }
-        let LibraryMapFactoryResult::JavaFactory { kind } = contract.result else {
-            return None;
-        };
-        match kind {
-            JavaMapFactoryKind::Of => {
-                let entries = &kids[1..];
-                if entries.len() % 2 != 0 {
-                    return None;
-                }
-                let values: Vec<ValueId> = entries.iter().map(|&kid| self.eval(kid, env)).collect();
-                let mut canonical_entries = Vec::with_capacity(values.len() / 2);
-                for kv in values.chunks(2) {
-                    canonical_entries.push(self.mk(ValOp::Seq(SEQ_VALUE_PAIR), kv.to_vec()));
-                }
-                Some(self.mk(ValOp::Seq(SEQ_VALUE_MAP), canonical_entries))
-            }
-            JavaMapFactoryKind::OfEntries => {
-                let mut canonical_entries = Vec::with_capacity(kids.len().saturating_sub(1));
-                for &entry in &kids[1..] {
-                    let kv = self.eval_java_map_entry_pair_expr(entry, env)?;
-                    canonical_entries.push(self.mk(ValOp::Seq(SEQ_VALUE_PAIR), kv));
-                }
-                Some(self.mk(ValOp::Seq(SEQ_VALUE_MAP), canonical_entries))
-            }
-        }
-    }
-
-    fn eval_java_map_entry_pair_expr(
-        &mut self,
-        expr: NodeId,
-        env: &FxHashMap<u32, ValueId>,
-    ) -> Option<Vec<ValueId>> {
-        if self.il.kind(expr) != NodeKind::Call {
-            return None;
-        }
-        let kids = self.il.children(expr);
-        if kids.len() != 3 || self.il.kind(kids[0]) != NodeKind::Field {
-            return None;
-        }
-        let Payload::Name(method) = self.il.node(kids[0]).payload else {
-            return None;
-        };
-        let contract = library_java_map_entry_contract_by_hash(
-            self.il.meta.lang,
-            "Map",
-            self.interner.symbol_hash(method),
-        )?;
-        let LibraryApiCalleeContract::JavaUtilStaticMember { .. } = contract.callee else {
-            return None;
-        };
-        match library_api_contract_evidence_for_call(
-            self.il,
-            self.interner,
-            expr,
-            contract.id,
-            contract.callee,
-            2,
-        ) {
-            LibraryApiEvidenceStatus::Admitted => {}
-            LibraryApiEvidenceStatus::Rejected => return None,
-            LibraryApiEvidenceStatus::Missing => return None,
-        }
-        Some(vec![self.eval(kids[1], env), self.eval(kids[2], env)])
-    }
-
-    fn proven_java_map_entry_pair(&self, value: ValueId) -> Option<Vec<ValueId>> {
-        let node = &self.nodes[value as usize];
-        if !matches!(node.op, ValOp::Call(0)) || node.args.len() != 3 {
-            return None;
-        }
-        let args = node.args.clone();
-        let callee = &self.nodes[args[0] as usize];
-        let ValOp::Field(method) = callee.op else {
-            return None;
-        };
-        let contract = library_java_map_entry_contract_by_hash(self.il.meta.lang, "Map", method)?;
-        let LibraryApiCalleeContract::JavaUtilStaticMember { .. } = contract.callee else {
-            return None;
-        };
-        if callee.args.len() != 1 {
-            return None;
-        }
-        match self.library_api_evidence_for_value_call(
-            value,
-            args[0],
-            Some(callee.args[0]),
-            contract.id,
-            contract.callee,
-            2,
-        ) {
-            LibraryApiEvidenceStatus::Admitted => {}
-            LibraryApiEvidenceStatus::Rejected => return None,
-            LibraryApiEvidenceStatus::Missing => return None,
-        }
-        Some(args[1..].to_vec())
-    }
-
-    fn library_api_evidence_for_value_call(
-        &self,
-        value: ValueId,
-        callee: ValueId,
-        receiver: Option<ValueId>,
-        id: nose_semantics::LibraryApiContractId,
-        callee_contract: LibraryApiCalleeContract,
-        arg_count: usize,
-    ) -> LibraryApiEvidenceStatus {
-        library_api_contract_evidence_at_call_span(
-            self.il,
-            self.interner,
-            LibraryApiSpanEvidenceQuery {
-                call_span: self.node_span[value as usize],
-                callee_span: self.library_api_value_span(callee),
-                receiver_span: self.library_api_receiver_query_span(value, callee, receiver),
-                id,
-                callee: callee_contract,
-                arg_count,
-            },
-        )
-    }
-
-    fn library_api_evidence_admitted(
-        &self,
-        call: NodeId,
-        id: nose_semantics::LibraryApiContractId,
-        callee: LibraryApiCalleeContract,
-        arg_count: usize,
-    ) -> bool {
-        matches!(
-            library_api_contract_evidence_for_call(
-                self.il,
-                self.interner,
-                call,
-                id,
-                callee,
-                arg_count,
-            ),
-            LibraryApiEvidenceStatus::Admitted
-        )
-    }
-
-    fn library_api_value_span(&self, value: ValueId) -> Option<Span> {
-        match self.nodes[value as usize].op {
-            ValOp::ImportBinding { .. } | ValOp::ImportNamespace { .. } => None,
-            _ => self.node_span[value as usize],
-        }
-    }
-
-    fn library_api_receiver_query_span(
-        &self,
-        value: ValueId,
-        callee: ValueId,
-        receiver: Option<ValueId>,
-    ) -> Option<Span> {
-        let receiver_span = receiver.and_then(|receiver| self.library_api_value_span(receiver))?;
-        let Some(call_span) = self.node_span[value as usize] else {
-            return Some(receiver_span);
-        };
-        let Some(callee_span) = self.library_api_value_span(callee) else {
-            return Some(receiver_span);
-        };
-        if self
-            .source_call_receiver_span(call_span, callee_span)
-            .is_some_and(|source_receiver_span| source_receiver_span != receiver_span)
-        {
-            None
-        } else {
-            Some(receiver_span)
-        }
-    }
-
-    fn source_call_receiver_span(&self, call_span: Span, callee_span: Span) -> Option<Span> {
-        self.il.nodes.iter().enumerate().find_map(|(idx, node)| {
-            if node.kind != NodeKind::Call || node.span != call_span {
-                return None;
-            }
-            let call = NodeId(idx as u32);
-            let callee = self.il.children(call).first().copied()?;
-            if self.il.node(callee).span != callee_span {
-                return None;
-            }
-            self.il
-                .children(callee)
-                .first()
-                .map(|&receiver| self.il.node(receiver).span)
-        })
-    }
-
-    fn eval_js_like_constructed_collection_or_map(
-        &mut self,
-        expr: NodeId,
-        kids: &[NodeId],
-        env: &FxHashMap<u32, ValueId>,
-    ) -> Option<ValueId> {
-        if !construct_syntax_proof(self.il, expr)
-            || kids.len() != 2
-            || self.il.kind(kids[0]) != NodeKind::Var
-        {
-            return None;
-        }
-        let Payload::Name(name) = self.il.node(kids[0]).payload else {
-            return None;
-        };
-        let constructor = self.interner.resolve(name);
-        if let Some(contract) =
-            library_js_like_set_constructor_contract(self.il.meta.lang, constructor)
-        {
-            let LibraryApiCalleeContract::JsGlobalConstructor { .. } = contract.callee else {
-                return None;
-            };
-            match library_api_contract_evidence_for_call(
-                self.il,
-                self.interner,
-                expr,
-                contract.id,
-                contract.callee,
-                1,
-            ) {
-                LibraryApiEvidenceStatus::Admitted => {}
-                LibraryApiEvidenceStatus::Rejected => return None,
-                LibraryApiEvidenceStatus::Missing => return None,
-            }
-            if !self.is_static_non_float_collection_expr(kids[1]) {
-                return None;
-            }
-            return Some(self.eval_membership_collection(kids[1], env));
-        }
-        let contract = library_js_like_map_constructor_contract(self.il.meta.lang, constructor)?;
-        let LibraryApiCalleeContract::JsGlobalConstructor { .. } = contract.callee else {
-            return None;
-        };
-        match library_api_contract_evidence_for_call(
-            self.il,
-            self.interner,
-            expr,
-            contract.id,
-            contract.callee,
-            1,
-        ) {
-            LibraryApiEvidenceStatus::Admitted => {}
-            LibraryApiEvidenceStatus::Rejected => return None,
-            LibraryApiEvidenceStatus::Missing => return None,
-        }
-        let LibraryMapFactoryResult::EntrySequence { entry_seq_tag } = contract.result else {
-            return None;
-        };
-        let entries = self.eval(kids[1], env);
-        self.map_factory_from_seq(entries, entry_seq_tag)
-    }
-
-    fn proven_go_literal_zero_map_value(&self, value: ValueId) -> Option<(ValueId, ValueId)> {
-        let contract = go_zero_map_lookup_contract(self.il.meta.lang)?;
-        let node = &self.nodes[value as usize];
-        if !matches!(node.op, ValOp::Seq(tag) if tag == stable_symbol_hash(contract.canonical_value_tag))
-            || node.args.len() != 2
-        {
-            return None;
-        }
-        Some((node.args[1], node.args[0]))
-    }
-
-    fn proven_go_literal_zero_map_seq(
-        &mut self,
-        expr: NodeId,
-        args: &[ValueId],
-    ) -> Option<ValueId> {
-        let contract = go_zero_map_literal_contract_for_node(self.il, self.interner, expr)?;
-        if args.is_empty() {
-            return None;
-        }
-        let entry_nodes = self.il.children(expr).to_vec();
-        if entry_nodes.len() != args.len() {
-            return None;
-        }
-        let mut canonical_entries = Vec::with_capacity(args.len());
-        let mut value_kind = None;
-        let mut default = None;
-        for (&entry_node_id, &entry_value) in entry_nodes.iter().zip(args.iter()) {
-            go_zero_map_entry_contract_for_node(self.il, self.interner, entry_node_id)?;
-            let kv_nodes = self.il.children(entry_node_id);
-            if kv_nodes.len() != 2
-                || !matches!(self.il.node(kv_nodes[0]).payload, Payload::LitStr(_))
-            {
-                return None;
-            }
-            let kind =
-                go_zero_map_default_kind(self.il.meta.lang, self.il.node(kv_nodes[1]).payload)?;
-            let value_default = self.go_literal_zero_default_value(kind);
-            match value_kind {
-                Some(current_kind) if current_kind != kind => return None,
-                Some(_) => {}
-                None => {
-                    value_kind = Some(kind);
-                    default = Some(value_default);
-                }
-            }
-            let entry_value_node = &self.nodes[entry_value as usize];
-            if !matches!(entry_value_node.op, ValOp::Seq(tag) if tag == stable_symbol_hash(contract.entry_tag))
-                || entry_value_node.args.len() != 2
-            {
-                return None;
-            }
-            canonical_entries
-                .push(self.mk(ValOp::Seq(SEQ_VALUE_PAIR), entry_value_node.args.clone()));
-        }
-        let map = self.mk(ValOp::Seq(SEQ_VALUE_MAP), canonical_entries);
-        Some(self.mk(
-            ValOp::Seq(stable_symbol_hash(contract.canonical_value_tag)),
-            vec![default?, map],
-        ))
-    }
-
-    fn go_literal_zero_default_value(&mut self, kind: GoZeroMapDefaultKind) -> ValueId {
-        match kind {
-            GoZeroMapDefaultKind::Int => self.int_const(0),
-            GoZeroMapDefaultKind::String => {
-                self.mk(ValOp::Const(stable_string_const_key("")), vec![])
-            }
-            GoZeroMapDefaultKind::Bool => self.mk(ValOp::Const(0x3000_0001), vec![]),
-            GoZeroMapDefaultKind::Float => {
-                self.mk(ValOp::Const(stable_float_const_key("0.0")), vec![])
-            }
-            GoZeroMapDefaultKind::Null => self.null_const(),
-        }
-    }
-
-    fn proven_map_value(&mut self, value: ValueId) -> Option<ValueId> {
-        if matches!(self.nodes[value as usize].op, ValOp::Seq(SEQ_VALUE_MAP)) {
-            return Some(value);
-        }
-        self.proven_free_name_map_factory(value)
-            .or_else(|| self.proven_java_map_factory_entries(value))
-            .or_else(|| {
-                self.proven_go_literal_zero_map_value(value)
-                    .map(|(map, _)| map)
-            })
-    }
-
-    fn proven_map_get_value(&mut self, value: ValueId) -> Option<(ValueId, ValueId)> {
-        let node = &self.nodes[value as usize];
-        if !matches!(node.op, ValOp::Call(0)) || node.args.len() != 2 {
-            return None;
-        }
-        let args = node.args.clone();
-        let callee = &self.nodes[node.args[0] as usize];
-        let ValOp::Field(method) = callee.op else {
-            return None;
-        };
-        let contract = library_map_get_contract_by_hash(
-            self.il.meta.lang,
-            method,
-            args.len().saturating_sub(1),
-        )?;
-        if callee.args.len() != 1 {
-            return None;
-        }
-        match self.library_api_evidence_for_value_call(
-            value,
-            args[0],
-            Some(callee.args[0]),
-            contract.id,
-            contract.callee,
-            args.len().saturating_sub(1),
-        ) {
-            LibraryApiEvidenceStatus::Admitted => {}
-            LibraryApiEvidenceStatus::Rejected => return None,
-            LibraryApiEvidenceStatus::Missing => return None,
-        }
-        let map = callee.args[0];
-        let map = if self.is_param_value(map, DomainEvidence::Map) {
-            map
-        } else {
-            self.proven_map_value(map)?
-        };
-        Some((map, args[1]))
-    }
-
-    fn proven_map_key_view_value(&mut self, value: ValueId) -> Option<ValueId> {
-        self.proven_map_key_view_value_matching(value, MapKeyViewKind::Collection)
-    }
-
-    fn proven_map_key_view_value_matching(
-        &mut self,
-        value: ValueId,
-        accepted: MapKeyViewKind,
-    ) -> Option<ValueId> {
-        let node = &self.nodes[value as usize];
-        if !matches!(node.op, ValOp::Call(0)) {
-            return None;
-        }
-        let args = node.args.clone();
-        if args.len() == 1 {
-            let callee = &self.nodes[args[0] as usize];
-            let ValOp::Field(method) = callee.op else {
-                return None;
-            };
-            let contract = library_map_key_view_contract_by_hash(self.il.meta.lang, method, 0)?;
-            if contract.result.kind != accepted || callee.args.len() != 1 {
-                return None;
-            }
-            match self.library_api_evidence_for_value_call(
-                value,
-                args[0],
-                Some(callee.args[0]),
-                contract.id,
-                contract.callee,
-                0,
-            ) {
-                LibraryApiEvidenceStatus::Admitted => {}
-                LibraryApiEvidenceStatus::Rejected => return None,
-                LibraryApiEvidenceStatus::Missing => return None,
-            }
-            let map = callee.args[0];
-            return if self.is_param_value(map, DomainEvidence::Map) {
-                Some(map)
-            } else {
-                self.proven_map_value(map)
-            };
-        }
-        if args.len() == 2 {
-            let callee = &self.nodes[args[0] as usize];
-            let ValOp::Field(method) = callee.op else {
-                return None;
-            };
-            let contract = library_map_key_view_wrapper_contract_by_hash(
-                self.il.meta.lang,
-                "Array",
-                method,
-                1,
-            )?;
-            if accepted != MapKeyViewKind::Collection || callee.args.len() != 1 {
-                return None;
-            }
-            let receiver_span = callee
-                .args
-                .first()
-                .and_then(|&receiver| self.node_span[receiver as usize]);
-            match library_api_contract_evidence_at_call_span(
-                self.il,
-                self.interner,
-                LibraryApiSpanEvidenceQuery {
-                    call_span: self.node_span[value as usize],
-                    callee_span: self.node_span[args[0] as usize],
-                    receiver_span,
-                    id: contract.id,
-                    callee: contract.callee,
-                    arg_count: 1,
-                },
-            ) {
-                LibraryApiEvidenceStatus::Admitted => {}
-                LibraryApiEvidenceStatus::Rejected => return None,
-                LibraryApiEvidenceStatus::Missing => return None,
-            }
-            return self.proven_map_key_view_value_matching(args[1], MapKeyViewKind::Iterator);
-        }
-        None
-    }
-
-    fn proven_map_key_view_expr(
-        &mut self,
-        expr: NodeId,
-        env: &FxHashMap<u32, ValueId>,
-    ) -> Option<ValueId> {
-        let value = self.eval(expr, env);
-        self.proven_map_key_view_value(value)
-    }
-
-    fn eval_proven_collection_membership_call(
-        &mut self,
-        expr: NodeId,
-        kids: &[NodeId],
-        env: &FxHashMap<u32, ValueId>,
-    ) -> Option<ValueId> {
-        let &callee = kids.first()?;
-        let Payload::Name(name) = self.il.node(callee).payload else {
-            return None;
-        };
-        let method = self.interner.resolve(name);
-        if self.il.kind(callee) != NodeKind::Field {
-            return None;
-        }
-        let callee_kids = self.il.children(callee);
-        let receiver = callee_kids.first().copied();
-
-        let contract =
-            library_method_call_contract(self.il.meta.lang, method, kids.len().saturating_sub(1))?;
-        if !self.library_api_evidence_admitted(
-            expr,
-            contract.id,
-            contract.callee,
-            kids.len().saturating_sub(1),
-        ) {
-            return None;
-        }
-        let result = contract.result;
-        if result.semantic != MethodSemanticContract::Builtin(Builtin::Contains) {
-            return None;
-        }
-
-        if result.args == MethodBuiltinArgs::FirstThenReceiver
-            && matches!(
-                result.receiver,
-                MethodReceiverContract::ExactCollection
-                    | MethodReceiverContract::ExactCollectionOrMap
-                    | MethodReceiverContract::ExactCollectionOrJavaKeySet
-                    | MethodReceiverContract::ExactSetOrMap
-            )
-            && kids.len() == 2
-        {
-            let receiver = receiver?;
-            let element = self.eval(kids[1], env);
-            if matches!(
-                result.receiver,
-                MethodReceiverContract::ExactCollectionOrMap
-                    | MethodReceiverContract::ExactCollectionOrJavaKeySet
-            ) {
-                if let Some(map) = self.proven_map_key_view_expr(receiver, env) {
-                    return Some(self.mk(ValOp::Bin(Op::In as u32), vec![element, map]));
-                }
-            }
-            let receiver_value = self.eval(receiver, env);
-            if let Some(collection) = self
-                .proven_collection_value(receiver_value)
-                .or_else(|| self.proven_local_collection_binding_value(receiver, env))
-            {
-                let collection = self.canonical_membership_collection_value(collection);
-                return Some(self.mk(ValOp::Bin(Op::In as u32), vec![element, collection]));
-            }
-            let receiver_param_safe = match result.receiver {
-                MethodReceiverContract::ExactSetOrMap => self.is_set_param_expr(receiver),
-                _ => self.is_collection_param_expr(receiver),
-            };
-            if receiver_param_safe {
-                let collection = self.eval_membership_collection(receiver, env);
-                return Some(self.mk(ValOp::Bin(Op::In as u32), vec![element, collection]));
-            }
-            None
-        } else if result.args == MethodBuiltinArgs::GoSliceContains && kids.len() == 3 {
-            let receiver = receiver?;
-            let MethodReceiverContract::ImportedNamespace(module) = result.receiver else {
-                return None;
-            };
-            if !self.is_import_namespace_expr(receiver, module, env)
-                && !self.file_imports_namespace(receiver, module)
-            {
-                return None;
-            }
-            let element = self.eval(kids[2], env);
-            let collection = if self.is_collection_param_expr(kids[1]) {
-                self.eval_membership_collection(kids[1], env)
-            } else {
-                self.proven_collection_expr(kids[1], env)?
-            };
-            let collection = self.canonical_membership_collection_value(collection);
-            Some(self.mk(ValOp::Bin(Op::In as u32), vec![element, collection]))
-        } else {
-            None
-        }
-    }
-
-    fn eval_proven_map_key_membership_call(
-        &mut self,
-        expr: NodeId,
-        kids: &[NodeId],
-        env: &FxHashMap<u32, ValueId>,
-    ) -> Option<ValueId> {
-        if kids.len() != 2 {
-            return None;
-        }
-        let callee = kids[0];
-        if self.il.kind(callee) != NodeKind::Field {
-            return None;
-        }
-        let Payload::Name(name) = self.il.node(callee).payload else {
-            return None;
-        };
-        let method = self.interner.resolve(name);
-        let contract = library_method_call_contract(self.il.meta.lang, method, 1)?;
-        if !self.library_api_evidence_admitted(expr, contract.id, contract.callee, 1) {
-            return None;
-        }
-        let result = contract.result;
-        if result.semantic != MethodSemanticContract::Builtin(Builtin::Contains)
-            || result.args != MethodBuiltinArgs::FirstThenReceiver
-            || !matches!(
-                result.receiver,
-                MethodReceiverContract::ExactMap
-                    | MethodReceiverContract::ExactCollectionOrMap
-                    | MethodReceiverContract::ExactSetOrMap
-            )
-        {
-            return None;
-        }
-        let receiver = self.il.children(callee).first().copied()?;
-        let key = self.eval(kids[1], env);
-        let receiver_value = self.eval(receiver, env);
-        let map = if self.is_map_param_expr(receiver) {
-            receiver_value
-        } else {
-            self.proven_map_value(receiver_value)
-                .or_else(|| self.proven_local_map_binding_value(receiver, env))?
-        };
-        Some(self.mk(ValOp::Bin(Op::In as u32), vec![key, map]))
-    }
-
-    fn eval_proven_map_get_default_call(
-        &mut self,
-        expr: NodeId,
-        kids: &[NodeId],
-        env: &FxHashMap<u32, ValueId>,
-    ) -> Option<ValueId> {
-        if kids.len() != 3 {
-            return None;
-        }
-        let callee = kids[0];
-        if self.il.kind(callee) != NodeKind::Field {
-            return None;
-        }
-        let Payload::Name(name) = self.il.node(callee).payload else {
-            return None;
-        };
-        let contract = library_method_call_contract(
-            self.il.meta.lang,
-            self.interner.resolve(name),
-            kids.len().saturating_sub(1),
-        )?;
-        if !self.library_api_evidence_admitted(
-            expr,
-            contract.id,
-            contract.callee,
-            kids.len().saturating_sub(1),
-        ) {
-            return None;
-        }
-        let result = contract.result;
-        if result.semantic != MethodSemanticContract::Builtin(Builtin::GetOrDefault)
-            || result.receiver != MethodReceiverContract::ExactMap
-            || !matches!(
-                result.args,
-                MethodBuiltinArgs::MapGetDefault | MethodBuiltinArgs::MapGetDefaultOrZeroArgLambda
-            )
-        {
-            return None;
-        }
-        let receiver = self.il.children(callee).first().copied()?;
-        let receiver_value = self.eval(receiver, env);
-        let map = if self.is_map_param_expr(receiver) {
-            receiver_value
-        } else {
-            self.proven_map_value(receiver_value)
-                .or_else(|| self.proven_local_map_binding_value(receiver, env))?
-        };
-        let key = self.eval(kids[1], env);
-        let default = self.eval_map_get_default_arg(result.args, kids[2], env)?;
-        Some(self.mk(
-            ValOp::Call(builtin_tag(Builtin::GetOrDefault)),
-            vec![map, key, default],
-        ))
-    }
-
-    fn eval_map_get_default_arg(
-        &mut self,
-        contract: MethodBuiltinArgs,
-        default: NodeId,
-        env: &FxHashMap<u32, ValueId>,
-    ) -> Option<ValueId> {
-        match contract {
-            MethodBuiltinArgs::MapGetDefault => Some(self.eval(default, env)),
-            MethodBuiltinArgs::MapGetDefaultOrZeroArgLambda => {
-                if self.il.kind(default) == NodeKind::Lambda {
-                    return self.eval_zero_arg_lambda_body(default, env);
-                }
-                Some(self.eval(default, env))
-            }
-            _ => None,
-        }
-    }
-
-    fn eval_zero_arg_lambda_body(
-        &mut self,
-        lambda: NodeId,
-        env: &FxHashMap<u32, ValueId>,
-    ) -> Option<ValueId> {
-        if self.il.kind(lambda) != NodeKind::Lambda {
-            return None;
-        }
-        let kids = self.il.children(lambda);
-        if kids.len() != 1 {
-            return None;
-        }
-        self.eval_lambda_body(lambda, &[], env)
-    }
-
-    fn eval_proven_integer_method_call(
-        &mut self,
-        call: NodeId,
-        kids: &[NodeId],
-        env: &FxHashMap<u32, ValueId>,
-    ) -> Option<ValueId> {
-        let &callee = kids.first()?;
-        if self.il.kind(callee) != NodeKind::Field {
-            return None;
-        }
-        let Payload::Name(name) = self.il.node(callee).payload else {
-            return None;
-        };
-        let method = self.interner.resolve(name);
-        let arg_count = kids.len().saturating_sub(1);
-        let contract =
-            library_scalar_integer_method_contract(self.il.meta.lang, method, arg_count)?;
-        if !matches!(
-            library_api_contract_evidence_for_call(
-                self.il,
-                self.interner,
-                call,
-                contract.id,
-                contract.callee,
-                arg_count,
-            ),
-            LibraryApiEvidenceStatus::Admitted
-        ) {
-            return None;
-        }
-        if contract.result.receiver != MethodReceiverContract::ExactInteger {
-            return None;
-        }
-        let receiver = self.il.children(callee).first().copied()?;
-        let receiver_value = self.eval_proven_integer_expr(receiver, env)?;
-        match contract.result.semantic {
-            ScalarIntegerMethod::Abs => Some(self.mk(ValOp::Un(ABS_CODE), vec![receiver_value])),
-            ScalarIntegerMethod::Min => {
-                let rhs = self.eval(*kids.get(1)?, env);
-                self.eval_proven_integer_minmax_method_call(MIN_CODE, receiver_value, rhs)
-            }
-            ScalarIntegerMethod::Max => {
-                let rhs = self.eval(*kids.get(1)?, env);
-                self.eval_proven_integer_minmax_method_call(MAX_CODE, receiver_value, rhs)
-            }
-            ScalarIntegerMethod::Clamp => {
-                let lo = self.eval(*kids.get(1)?, env);
-                let hi = self.eval(*kids.get(2)?, env);
-                self.eval_proven_integer_clamp_method_call(receiver_value, lo, hi)
-            }
-        }
-    }
-
-    fn eval_proven_integer_minmax_method_call(
-        &mut self,
-        op: u32,
-        receiver: ValueId,
-        rhs: ValueId,
-    ) -> Option<ValueId> {
-        if !self.is_integer_domain_value(rhs) {
-            return None;
-        }
-        Some(self.mk(ValOp::Bin(op), vec![receiver, rhs]))
-    }
-
-    fn eval_proven_free_minmax_call(
-        &mut self,
-        expr: NodeId,
-        kids: &[NodeId],
-        env: &FxHashMap<u32, ValueId>,
-    ) -> Option<ValueId> {
-        if kids.len() != 3 || self.il.kind(kids[0]) != NodeKind::Var {
-            return None;
-        }
-        let Payload::Name(name) = self.il.node(kids[0]).payload else {
-            return None;
-        };
-        let method = self.interner.resolve(name);
-        let contract = library_free_function_builtin_contract(self.il.meta.lang, method, 2)?;
-        if !self.library_api_evidence_admitted(expr, contract.id, contract.callee, 2) {
-            return None;
-        }
-        let op = match (contract.result.builtin, contract.result.args) {
-            (Builtin::Min, BuiltinArgContract::All) => MIN_CODE,
-            (Builtin::Max, BuiltinArgContract::All) => MAX_CODE,
-            _ => return None,
-        };
-        let left = self.eval(kids[1], env);
-        let right = self.eval(kids[2], env);
-        Some(self.mk(ValOp::Bin(op), vec![left, right]))
-    }
-
-    fn eval_proven_integer_clamp_method_call(
-        &mut self,
-        receiver: ValueId,
-        lo: ValueId,
-        hi: ValueId,
-    ) -> Option<ValueId> {
-        if !self.is_integer_domain_value(lo) || !self.is_integer_domain_value(hi) {
-            return None;
-        }
-        self.proof_backed_clamp_value(receiver, lo, hi)
-    }
-
-    fn eval_proven_integer_expr(
-        &mut self,
-        expr: NodeId,
-        env: &FxHashMap<u32, ValueId>,
-    ) -> Option<ValueId> {
-        let value = self.eval(expr, env);
-        if self.il.kind(expr) == NodeKind::Var {
-            return self.is_integer_param_expr(expr).then_some(value);
-        }
-        self.is_integer_domain_value(value).then_some(value)
-    }
-
-    fn is_integer_domain_value(&self, value: ValueId) -> bool {
-        if self.int_const_value(value).is_some()
-            || self.is_param_value(value, DomainEvidence::Integer)
-        {
-            return true;
-        }
-        let node = &self.nodes[value as usize];
-        match node.op {
-            ValOp::Un(op) if op == ABS_CODE && node.args.len() == 1 => {
-                self.is_integer_domain_value(node.args[0])
-            }
-            ValOp::Bin(op) if (op == MIN_CODE || op == MAX_CODE) && node.args.len() == 2 => node
-                .args
-                .iter()
-                .copied()
-                .all(|arg| self.is_integer_domain_value(arg)),
-            ValOp::Clamp if node.args.len() == 3 => node
-                .args
-                .iter()
-                .copied()
-                .all(|arg| self.is_integer_domain_value(arg)),
-            _ => false,
-        }
     }
 
     /// Push an effect sink, tagged with the current path condition — so a *conditional*
@@ -3072,250 +1521,6 @@ impl<'a> Builder<'a> {
             }
         }
         self.flush_fields();
-    }
-
-    fn seed_immutable_bindings(&mut self, root: NodeId, context: Option<&ValueFingerprintContext>) {
-        if let Some(context) = context {
-            let required = context.module.required_bindings_for(self.il, root);
-            self.seed_module_value_bindings_from_context(&context.module, Some(&required));
-        } else {
-            self.seed_module_value_bindings();
-        }
-        if let Some(context) = context {
-            self.seed_function_binding_hashes(&context.function_bindings);
-        } else {
-            self.seed_function_bindings();
-        }
-    }
-
-    fn seed_module_value_bindings(&mut self) {
-        let mut counts: FxHashMap<Symbol, usize> = FxHashMap::default();
-        for stmt in self.top_level_statements() {
-            let Some(name) = self.assignment_name(stmt) else {
-                continue;
-            };
-            *counts.entry(name).or_insert(0) += 1;
-        }
-
-        let top_level = self.top_level_statements();
-        self.seed_module_value_bindings_from_parts(&top_level, &counts, None, None, None);
-    }
-
-    fn seed_module_value_bindings_from_context(
-        &mut self,
-        context: &ModuleSeedContext,
-        required_bindings: Option<&FxHashSet<Symbol>>,
-    ) {
-        self.seed_module_value_bindings_from_parts(
-            &context.top_level,
-            &context.assignment_counts,
-            Some(&context.mutated_bindings),
-            Some(&context.unit_symbols),
-            required_bindings,
-        );
-    }
-
-    fn seed_module_value_bindings_from_parts(
-        &mut self,
-        top_level: &[NodeId],
-        counts: &FxHashMap<Symbol, usize>,
-        mutated_bindings: Option<&FxHashSet<Symbol>>,
-        unit_symbols: Option<&FxHashSet<Symbol>>,
-        required_bindings: Option<&FxHashSet<Symbol>>,
-    ) {
-        let mut env: FxHashMap<u32, ValueId> = FxHashMap::default();
-        for &stmt in top_level {
-            let kids = self.il.children(stmt);
-            if kids.len() != 2 {
-                continue;
-            }
-            let Some(name) = self.assignment_name(stmt) else {
-                continue;
-            };
-            if required_bindings.is_some_and(|required| !required.contains(&name)) {
-                continue;
-            }
-            let unit_defines_symbol = unit_symbols
-                .map(|symbols| symbols.contains(&name))
-                .unwrap_or_else(|| self.unit_defines_symbol(name));
-            if unit_defines_symbol {
-                continue;
-            }
-            if counts.get(&name).copied().unwrap_or(0) != 1 {
-                continue;
-            }
-            let mutated = mutated_bindings
-                .map(|bindings| bindings.contains(&name))
-                .unwrap_or_else(|| self.module_binding_mutated(name));
-            if mutated {
-                continue;
-            }
-            let value = self.eval(kids[1], &env);
-            let binding_domain =
-                nose_semantics::domain_evidence_for_binding_lhs(self.il, self.interner, kids[0]);
-            let value = if self.immutable_binding_safe(kids[1], &env) {
-                value
-            } else if binding_domain.is_some_and(|domain| domain.is_map()) {
-                let Some(proven) = self.proven_map_value(value) else {
-                    continue;
-                };
-                proven
-            } else if binding_domain.is_some_and(|domain| domain.is_collection_or_set()) {
-                let Some(proven) = self.proven_collection_value(value) else {
-                    continue;
-                };
-                proven
-            } else {
-                continue;
-            };
-            if let Payload::Cid(cid) = self.il.node(kids[0]).payload {
-                env.insert(cid, value);
-            }
-            self.global_env.insert(name, value);
-        }
-    }
-
-    fn top_level_statements(&self) -> Vec<NodeId> {
-        top_level_statements_for(self.il)
-    }
-
-    fn seed_function_bindings(&mut self) {
-        let bindings = self.collect_function_binding_hashes();
-        self.seed_function_binding_hashes(&bindings);
-    }
-
-    fn collect_function_binding_hashes(&mut self) -> Vec<(Symbol, u64)> {
-        let mut bindings = Vec::new();
-        for unit in self.il.units.clone() {
-            if !matches!(unit.kind, UnitKind::Function | UnitKind::Method) {
-                continue;
-            }
-            let Some(name) = unit.name else {
-                continue;
-            };
-            if self.function_binding_safe(unit.root, unit.root) {
-                let hash = self.valued_subtree_hash(unit.root);
-                bindings.push((name, hash));
-            }
-        }
-        bindings
-    }
-
-    fn seed_function_binding_hashes(&mut self, bindings: &[(Symbol, u64)]) {
-        for &(name, hash) in bindings {
-            let value = self.mk(ValOp::Lambda(hash), vec![]);
-            self.global_env.insert(name, value);
-        }
-    }
-
-    fn assignment_name(&self, stmt: NodeId) -> Option<Symbol> {
-        module_seed_assignment_name(self.il, stmt, &self.local_scope_nodes)
-    }
-
-    fn unit_defines_symbol(&self, symbol: Symbol) -> bool {
-        self.il
-            .units
-            .iter()
-            .any(|unit| unit.name.is_some_and(|name| name == symbol))
-    }
-
-    fn module_binding_mutated(&self, name: Symbol) -> bool {
-        let top_level = self.top_level_statements();
-        let shadowed = shadowed_js_like_module_binding_nodes_for_symbol_in_scope(
-            self.il,
-            name,
-            &self.local_scope_nodes,
-        );
-        self.il.nodes.iter().enumerate().any(|(idx, node)| {
-            let node_id = NodeId(idx as u32);
-            if shadowed.contains(&node_id) {
-                return false;
-            }
-            match node.kind {
-                NodeKind::Call => self.call_mutates_binding(node_id, name).unwrap_or(false),
-                NodeKind::Assign if !top_level.contains(&node_id) => self
-                    .assignment_mutates_binding(node_id, name)
-                    .unwrap_or(false),
-                _ => false,
-            }
-        })
-    }
-
-    fn assignment_mutates_binding(&self, assign: NodeId, name: Symbol) -> Option<bool> {
-        let lhs = binding_write_target(self.il, assign)?;
-        Some(self.node_contains_symbol(lhs, name))
-    }
-
-    fn call_mutates_binding(&self, call: NodeId, name: Symbol) -> Option<bool> {
-        if let Some(receiver) = receiver_mutation_call_receiver(self.il, self.interner, call) {
-            return Some(self.node_refers_to_symbol(receiver, name));
-        }
-        if let Some(args) = opaque_argument_escape_args(self.il, call) {
-            return Some(args.iter().any(|&arg| self.node_contains_symbol(arg, name)));
-        }
-        Some(false)
-    }
-
-    fn node_refers_to_symbol(&self, node: NodeId, name: Symbol) -> bool {
-        self.node_symbol(node).is_some_and(|symbol| symbol == name)
-    }
-
-    fn node_symbol(&self, node: NodeId) -> Option<Symbol> {
-        node_symbol_in_scope(self.il, node, &self.local_scope_nodes)
-    }
-
-    fn node_contains_symbol(&self, node: NodeId, name: Symbol) -> bool {
-        self.node_refers_to_symbol(node, name)
-            || self
-                .il
-                .children(node)
-                .iter()
-                .any(|&child| self.node_contains_symbol(child, name))
-    }
-
-    fn node_refers_to_cid(&self, node: NodeId, cid: u32) -> bool {
-        matches!(self.il.node(node).payload, Payload::Cid(current) if current == cid)
-    }
-
-    fn node_contains_cid(&self, node: NodeId, cid: u32) -> bool {
-        self.node_refers_to_cid(node, cid)
-            || self
-                .il
-                .children(node)
-                .iter()
-                .any(|&child| self.node_contains_cid(child, cid))
-    }
-
-    fn immutable_binding_safe(&self, node: NodeId, env: &FxHashMap<u32, ValueId>) -> bool {
-        match self.il.kind(node) {
-            NodeKind::Raw
-            | NodeKind::Call
-            | NodeKind::HoF
-            | NodeKind::Func
-            | NodeKind::Lambda
-            | NodeKind::Loop
-            | NodeKind::Try
-            | NodeKind::Throw
-            | NodeKind::Assign => false,
-            NodeKind::Var => match self.il.node(node).payload {
-                Payload::Cid(c) => env.contains_key(&c),
-                Payload::Name(s) => self.global_env.contains_key(&s),
-                _ => false,
-            },
-            NodeKind::Lit => matches!(
-                self.il.node(node).payload,
-                Payload::LitInt(_)
-                    | Payload::LitBool(_)
-                    | Payload::LitStr(_)
-                    | Payload::LitFloat(_)
-                    | Payload::Lit(nose_il::LitClass::Null)
-            ),
-            _ => self
-                .il
-                .children(node)
-                .iter()
-                .all(|&c| self.immutable_binding_safe(c, env)),
-        }
     }
 
     /// Build the interprocedural inline registry (see the `inline_fns` field): pure,
@@ -8755,326 +6960,6 @@ impl<'a> Builder<'a> {
     fn seq_surface(&self, node: NodeId) -> Option<SeqSurfaceContract> {
         seq_surface_contract_for_node(self.il, self.interner, node)
     }
-}
-
-fn collect_assigned(il: &Il, node: NodeId, out: &mut FxHashSet<u32>) {
-    if il.kind(node) == NodeKind::Assign {
-        if let Some(&lhs) = il.children(node).first() {
-            if il.kind(lhs) == NodeKind::Var {
-                if let Payload::Cid(c) = il.node(lhs).payload {
-                    out.insert(c);
-                }
-            }
-        }
-    }
-    for &c in il.children(node) {
-        collect_assigned(il, c, out);
-    }
-}
-
-fn op_code(p: Payload) -> u32 {
-    match p {
-        Payload::Op(op) => op as u32,
-        _ => 0,
-    }
-}
-
-/// All canonical variable ids referenced anywhere in `node`'s subtree.
-fn mentioned_cids(il: &Il, node: NodeId) -> FxHashSet<u32> {
-    let mut out = FxHashSet::default();
-    mentioned_scan(il, node, &mut out);
-    out
-}
-
-fn mentioned_scan(il: &Il, node: NodeId, out: &mut FxHashSet<u32>) {
-    if il.kind(node) == NodeKind::Var {
-        if let Payload::Cid(c) = il.node(node).payload {
-            out.insert(c);
-        }
-    }
-    for &c in il.children(node) {
-        mentioned_scan(il, c, out);
-    }
-}
-
-/// Loop induction variables: those updated by `i = i ± constant` in the body.
-fn induction_vars(il: &Il, body: NodeId) -> FxHashSet<u32> {
-    let mut out = FxHashSet::default();
-    induction_scan(il, body, &mut out);
-    out
-}
-
-fn induction_scan(il: &Il, node: NodeId, out: &mut FxHashSet<u32>) {
-    if il.kind(node) == NodeKind::Assign {
-        let kids = il.children(node);
-        if kids.len() == 2 && il.kind(kids[0]) == NodeKind::Var {
-            if let Payload::Cid(c) = il.node(kids[0]).payload {
-                if is_increment(il, kids[1], c) {
-                    out.insert(c);
-                }
-            }
-        }
-    }
-    for &c in il.children(node) {
-        induction_scan(il, c, out);
-    }
-}
-
-/// The constant step of induction variable `cid` if the body updates it *exactly once*
-/// as `i = i + k` / `i = k + i` / `i = i - k` (k an int literal); else `None`. `k - i`
-/// is a reflection (not a step) and is rejected, as is a variable updated 0 or ≥2 times.
-fn induction_step(il: &Il, body: NodeId, cid: u32) -> Option<i64> {
-    let mut step = None;
-    let mut count = 0u32;
-    induction_step_scan(il, body, cid, &mut step, &mut count);
-    if count == 1 {
-        step
-    } else {
-        None
-    }
-}
-
-fn induction_step_scan(il: &Il, node: NodeId, cid: u32, step: &mut Option<i64>, count: &mut u32) {
-    if il.kind(node) == NodeKind::Assign {
-        let kids = il.children(node);
-        if kids.len() == 2 && il.kind(kids[0]) == NodeKind::Var {
-            if let Payload::Cid(c) = il.node(kids[0]).payload {
-                if c == cid {
-                    *count += 1;
-                    *step = increment_amount(il, kids[1], cid);
-                }
-            }
-        }
-    }
-    for &c in il.children(node) {
-        induction_step_scan(il, c, cid, step, count);
-    }
-}
-
-/// The signed step if `expr` is `cid + k`, `k + cid`, or `cid - k` (k an int literal);
-/// `k - cid` and anything else → `None`.
-fn increment_amount(il: &Il, expr: NodeId, cid: u32) -> Option<i64> {
-    if il.kind(expr) != NodeKind::BinOp {
-        return None;
-    }
-    let kids = il.children(expr);
-    if kids.len() != 2 {
-        return None;
-    }
-    let is_self = |n: NodeId| matches!((il.kind(n), il.node(n).payload), (NodeKind::Var, Payload::Cid(c)) if c == cid);
-    let lit = |n: NodeId| match il.node(n).payload {
-        Payload::LitInt(v) => Some(v),
-        _ => None,
-    };
-    match il.node(expr).payload {
-        Payload::Op(Op::Add) => {
-            if is_self(kids[0]) {
-                lit(kids[1])
-            } else if is_self(kids[1]) {
-                lit(kids[0])
-            } else {
-                None
-            }
-        }
-        // Only `i - k` is a step; `k - i` reflects.
-        Payload::Op(Op::Sub) if is_self(kids[0]) => lit(kids[1]).map(|v| -v),
-        _ => None,
-    }
-}
-
-/// Whether `expr` is `cid ± literal` — a step of the induction variable `cid`.
-fn is_increment(il: &Il, expr: NodeId, cid: u32) -> bool {
-    if il.kind(expr) != NodeKind::BinOp
-        || !matches!(
-            il.node(expr).payload,
-            Payload::Op(Op::Add) | Payload::Op(Op::Sub)
-        )
-    {
-        return false;
-    }
-    let mut refs_self = false;
-    let mut others_literal = true;
-    for &k in il.children(expr) {
-        match (il.kind(k), il.node(k).payload) {
-            (NodeKind::Var, Payload::Cid(c)) if c == cid => refs_self = true,
-            (NodeKind::Lit, _) => {}
-            _ => others_literal = false,
-        }
-    }
-    refs_self && others_literal
-}
-
-/// The complementary comparison op code, if `opc` is a comparison; else `None`.
-fn negate_cmp_code(lang: Lang, opc: u32) -> Option<u32> {
-    let op = op_from_code(opc)?;
-    semantics(lang)
-        .operators()
-        .comparison_complement(op)
-        .map(|contract| contract.output as u32)
-}
-
-/// The same comparison with operands swapped: `a < b` becomes `b > a`.
-fn reverse_cmp_code(lang: Lang, opc: u32) -> Option<u32> {
-    let op = op_from_code(opc)?;
-    semantics(lang)
-        .operators()
-        .comparison_reverse(op)
-        .map(|contract| contract.output as u32)
-}
-
-fn op_from_code(opc: u32) -> Option<Op> {
-    const OPS: &[Op] = &[
-        Op::Add,
-        Op::Sub,
-        Op::Mul,
-        Op::Div,
-        Op::Mod,
-        Op::Pow,
-        Op::Eq,
-        Op::Ne,
-        Op::Lt,
-        Op::Le,
-        Op::Gt,
-        Op::Ge,
-        Op::In,
-        Op::And,
-        Op::Or,
-        Op::Not,
-        Op::BitAnd,
-        Op::BitOr,
-        Op::BitXor,
-        Op::Shl,
-        Op::Shr,
-        Op::BitNot,
-        Op::Neg,
-        Op::Pos,
-    ];
-    OPS.iter().copied().find(|op| *op as u32 == opc)
-}
-
-fn is_commutative(opc: u32) -> bool {
-    is_assoc_comm_code(opc)
-        || opc == Op::Eq as u32
-        || opc == Op::Ne as u32
-        || opc == MIN_CODE
-        || opc == MAX_CODE
-}
-
-/// Coarse type of a `Const` value node from its key range (mirrors the `eval` Lit keys):
-/// int range → Num, string range → Str, bool range → Bool, small `LitClass` codes → their
-/// type; sentinels (⊥, void-return, break) → Unknown.
-/// Is this type a concatenation monoid (string/list) — where `+` is non-commutative?
-fn const_value_domain(k: u32) -> ValueDomain {
-    match k {
-        0x1000_0000..=0x1FFF_FFFF => ValueDomain::Number,
-        0x2000_0000..=0x2FFF_FFFF => ValueDomain::String,
-        0x3000_0001 | 0x3000_0002 => ValueDomain::Boolean,
-        0x4000_0000..=0x4FFF_FFFF => ValueDomain::Number, // retained float
-        0 | 1 => ValueDomain::Number,                     // LitClass::Int / Float
-        2 => ValueDomain::String,                         // LitClass::Str
-        3 => ValueDomain::Boolean,                        // LitClass::Bool
-        _ => ValueDomain::Unknown,
-    }
-}
-
-/// Associative *and* commutative operators (flatten-eligible).
-fn is_assoc_comm_code(opc: u32) -> bool {
-    // NOTE: logical `And`/`Or` are deliberately ABSENT — short-circuit value-and/or is
-    // associative but NOT commutative (`1 or 2` ≠ `2 or 1`; it returns the deciding
-    // operand's value). Treating them as commutative here swapped their operands by hash
-    // and silently merged `a or b` with `b or a` (a false merge the post-normalize oracle
-    // can't see). They are instead rewritten to the positional `Phi` form in `mk`.
-    opc == Op::Add as u32
-        || opc == Op::Mul as u32
-        || opc == Op::BitAnd as u32
-        || opc == Op::BitOr as u32
-        || opc == Op::BitXor as u32
-}
-
-/// `Reduce` op codes for the selection reductions (min/max). Kept clear of the small
-/// `Op` discriminants (used for `+`/`*` folds) and of the `Const` int range.
-const REDUCE_MAX: u32 = 0xFF00;
-const REDUCE_MIN: u32 = 0xFF01;
-/// `Reduce` op codes for the boolean short-circuit reductions: `any`/`some` (existential
-/// OR) and `all`/`every` (universal AND). `REDUCE_ALL == REDUCE_ANY + 1` (the fold reuses
-/// the offset to pick the OR/AND identity false/true).
-const REDUCE_ANY: u32 = 0xFF02;
-const REDUCE_ALL: u32 = 0xFF03;
-const ORDERED_STRING_JOIN: u32 = 0xFF04;
-
-/// `Un` op code for absolute value — `abs(x)` and the `x if x>=0 else -x` idiom both
-/// canonicalize to `Un(ABS_CODE, [x])`. Clear of the small `Op` discriminants.
-const ABS_CODE: u32 = 0xAB5;
-/// Pseudo-ops for the 2-way min/max idiom (`x if x<y else y` ≡ `min(x,y)`), clear of the
-/// `Op` discriminants and `ABS_CODE`. Commutative (min/max are symmetric).
-const MIN_CODE: u32 = 0x319;
-const MAX_CODE: u32 = 0x32A;
-const JS_PROTOTYPE_IN_CODE: u32 = 0x4A53_494E;
-const C_U16_BE_BYTE_PACK_CODE: u32 = 0x4331_3642;
-const C_U32_BE_BYTE_PACK_CODE: u32 = 0x4333_3242;
-const EFFECT_ORDINAL_SINK_TAG: u64 = 0xEFFE_C701;
-
-/// A selection reduction (min/max) keeps no additive/multiplicative identity, so its
-/// `Reduce` carries only the per-element contribution (no init) — a `max`-loop and
-/// `max(gen)` then converge regardless of the loop's incidental seed value.
-fn is_selection_code(op: u32) -> bool {
-    // any/all carry only the per-element predicate (no accumulator seed), like min/max.
-    op == REDUCE_MAX || op == REDUCE_MIN || op == REDUCE_ANY || op == REDUCE_ALL
-}
-
-/// The identity element of a reduction operator (`acc ⊕ identity = acc`), used to
-/// neutralize a filtered-out element in a guarded reduction. Only the operators with
-/// an integer-literal identity are handled (`+`→0, `*`→1).
-fn identity_of(opc: u32) -> Option<u32> {
-    if opc == Op::Add as u32 {
-        Some(0)
-    } else if opc == Op::Mul as u32 {
-        Some(1)
-    } else {
-        None
-    }
-}
-
-fn op_tag(op: &ValOp) -> u64 {
-    let (k, p): (u64, u64) = match op {
-        ValOp::Input(c) => (1, *c as u64),
-        ValOp::Const(c) => (2, *c as u64),
-        ValOp::Bin(o) => (3, *o as u64),
-        ValOp::Un(o) => (4, *o as u64),
-        ValOp::Field(n) => (5, *n),
-        ValOp::Index => (6, 0),
-        ValOp::Call(t) => (7, *t as u64),
-        ValOp::Hof(h) => (8, *h as u64),
-        ValOp::Clamp => (20, 0),
-        ValOp::Seq(t) => (9, *t),
-        ValOp::ImportNamespace { module_hash } => (21, *module_hash),
-        ValOp::ImportBinding {
-            module_hash,
-            exported_hash,
-        } => (22, combine(*module_hash, *exported_hash)),
-        ValOp::CollectionParam => (17, 0),
-        ValOp::ArrayParam => (18, 0),
-        ValOp::StringParam => (19, 0),
-        ValOp::Phi => (10, 0),
-        ValOp::Lambda(h) => (11, *h),
-        ValOp::Loop(c) => (12, *c as u64),
-        ValOp::Elem(h) => (14, *h),
-        ValOp::Reduce(o) => (15, *o as u64),
-        ValOp::Idx(h) => (16, *h),
-        ValOp::Formula(h) => (19, *h),
-        ValOp::Recurrence(h) => (18, *h),
-        ValOp::Opaque(c) => (13, *c),
-    };
-    combine(k.wrapping_mul(0xF00D), p)
-}
-
-fn stable_string_const_key(value: &str) -> u32 {
-    0x2000_0000u32.wrapping_add(stable_symbol_hash(value) as u32)
-}
-
-fn stable_float_const_key(value: &str) -> u32 {
-    let normalized = value.trim().trim_end_matches(['f', 'F', 'd', 'D']);
-    0x4000_0000u32.wrapping_add(stable_symbol_hash(normalized) as u32)
 }
 
 #[cfg(test)]
